isolated fetchScryfall function
