+++ src/app/editor/components/cards/mtgCardObject.tsx
@@ -7,7 +7,7 @@ import DOMPurify from "isomorphic-dompurify";
 import Image from "next/image";
 import {FaceType, MTGCard, MTGCardTemplate} from "@/lib/card";
 import {TbCaretUpDownFilled, TbCaretUpFilled} from "react-icons/tb";
-import {hasReverseFace, isolateFrontAndBackFaces} from "@/lib/mtg";
+import {hasReverseFace, isolateFrontAndBackFaces} from "@/lib/mtg/mtg";
 
 function getPowerToughnessText(power?: string, toughness?: string): string | null {
 	if ((power === undefined || power === "") && (toughness === undefined || toughness === "")) {
